requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==41.0.5
Django==4.2.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
djanko==0.1.1
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
